requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
asyncpg==0.29.0
black==24.10.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
fastapi==0.115.2
greenlet==3.1.1
idna==3.10
jwt==1.3.1
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
tzlocal==5.2
alembic~=1.13.3
sqlmodel~=0.0.22
pytest~=8.3.3
httpx~=0.27.2